+++ app/src/main/java/com/example/kaspukowaniev3/domain/usecase/RecipeDetailScreenUseCase/CalculateAmountOfBoxesUseCase.kt
@@ -13,8 +13,8 @@ class CalculateAmountOfBoxesUseCase @Inject constructor() {
         doseWeight: Double,
         capsulesGross: Double,
     ): String = if (isDataCorrect(boxWeight, weightOfPowder)) {
-        val amountOfFullCapsules = (weightOfPowder.toInt() / doseWeight) * capsulesGross
-        (amountOfFullCapsules / boxWeight.toInt())
+        val amountOfFullCapsules = (weightOfPowder.toDouble() / doseWeight) * capsulesGross
+        (amountOfFullCapsules / boxWeight.toDouble())
             .toBigDecimal()
             .setScale(0, RoundingMode.UP)
             .toString()
@@ -28,6 +28,6 @@ class CalculateAmountOfBoxesUseCase @Inject constructor() {
     ) =
         weightOfPowder.isNotBlank() &&
                 boxWeight.isNotBlank() &&
-                weightOfPowder.toInt() != 0 &&
-                boxWeight.toInt() != 0
+                weightOfPowder.toDouble() != 0.0 &&
+                boxWeight.toDouble() != 0.0
 }
